# PERF: reindex default fill_value dtype