--- a/pandas/core/internals/managers.py
+++ b/pandas/core/internals/managers.py
@@ -36,6 +36,7 @@ from pandas.core.dtypes.common import (
     ensure_platform_int,
     is_1d_only_ea_dtype,
     is_dtype_equal,
+    is_float_dtype,
     is_list_like,
 )
 from pandas.core.dtypes.dtypes import ExtensionDtype
@@ -817,6 +818,11 @@ class BaseBlockManager(DataManager):
 
         if fill_value is None:
             fill_value = np.nan
+            dtype = interleaved_dtype([blk.dtype for blk in self.blocks])
+            if is_float_dtype(dtype):
+                # GH45857 avoid unnecessary upcasting
+                fill_value = dtype.type(fill_value)
+
         block_shape = list(self.shape)
         block_shape[0] = len(placement)