@jbrockmendel 2022-06-17
@lukemanley do you happen to be using a windows machine or 32bit build locally?