@jbrockmendel 2022-06-17
There's a comment in NDFrame._reindex_with_indexers `# TODO: speed up on homogeneous DataFrame objects (see _reindex_multi)` that i think might be useful in these cases.  Could that be helpful?